pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm>=8", "wheel"]
build-backend = "setuptools.build_meta"


[project]
dynamic = ["version"]
name = "spirit-phantom"
description = "SPIRIT phantom tools"
readme = "README.md"
requires-python = ">=3.11,<3.14"
authors = [
  { name = 'Frank Bolton', email = 'frank.bolton@goldstandardphantoms.com' },
]
urls = { Homepage = "https://goldstandardphantoms.com" }


dependencies = [
  "numpy>=2.3.5",
  "itk-elastix>=0.23.0",
]


[dependency-groups]
dev = [
  "pytest~=8.1",
  "pytest-cov~=4.1",
  "tox-uv~=1.14",
  "mkdocs>=1.6.1",
  "mkdocs-material>=9.5.49",
  "mkdocstrings[python]>=0.27.0",
  "markdown-include>=0.8.1",
  "ruff>=0.14.3",
  "mypy>=1.18.2",
  "scipy-stubs>=1.16.3.2",
  # Scipy and nibabel included as development dependency as they are used
  # for testing the slice thickness function and registration.
  "scipy>=1.16.3",
  "nibabel>=5.3.3",
]

[project.scripts]
spirit-phantom = "spirit_phantom.cli:main"

[tool.setuptools]
package-data = { "spirit_phantom" = [
  "core/registration_parameters_Rigid.txt",
  "core/registration_parameters_Affine.txt",
] }
package-dir = { "" = "src" }
include-package-data = true

[tool.setuptools_scm]
